--- server/auth_service/src/auth/auth.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { AuthService } from './auth.service';
-
-@Controller()
-export class AuthController {
-  constructor(private readonly authService: AuthService) {}
-}
