--- server/auth_service/src/auth/auth.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-
-@Module({
-  controllers: [AuthController],
-  providers: [AuthService],
-})
-export class AuthModule {}
