+++ server/docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3.8'
+services:
+  users_db:
+    image: postgres:latest
+    environment:
+      POSTGRES_USER: users_admin
+      POSTGRES_PASSWORD: ${DB_ADMIN_PASSWORD}
+      POSTGRES_DB: users
+    volumes:
+      - users_db_data:/var/lib/postgresql/data
+      - ./users-service/init.sql:/docker-entrypoint-initdb.d/init.sql
+    ports:
+      - "5432:5432"
+
+volumes:
+  users_db_data:
