+++ server/users-service/init.sql
@@ -1,0 +1,16 @@
+GRANT ALL PRIVILEGES ON DATABASE users TO users_admin;
+
+\c users
+
+GRANT CREATE ON DATABASE users TO users_admin;
+GRANT CREATE ON SCHEMA public TO users_admin;
+
+GRANT ALL PRIVILEGES ON ALL TABLES IN SCHEMA public TO users_admin;
+GRANT ALL PRIVILEGES ON ALL SEQUENCES IN SCHEMA public TO users_admin;
+GRANT ALL PRIVILEGES ON ALL FUNCTIONS IN SCHEMA public TO users_admin;
+
+ALTER DEFAULT PRIVILEGES IN SCHEMA public GRANT ALL PRIVILEGES ON TABLES TO users_admin;
+ALTER DEFAULT PRIVILEGES IN SCHEMA public GRANT ALL PRIVILEGES ON SEQUENCES TO users_admin;
+ALTER DEFAULT PRIVILEGES IN SCHEMA public GRANT ALL PRIVILEGES ON FUNCTIONS TO users_admin;
+
+ALTER DATABASE users OWNER TO users_admin;
