+++ server/users-service/prisma/schema.prisma
@@ -1,0 +1,25 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model User {  
+  id          String   @id @default(cuid())
+  username    String   @unique @db.VarChar(24)
+  email       String   @unique @db.VarChar(24)
+  password    String
+  href        String   @unique
+  image       String?   
+  role        Role     @default(user)
+  createdAt   DateTime @default(now())
+  updatedAt   DateTime @updatedAt
+}
+
+enum Role {
+  user
+  admin
+}
