+++ server/users-service/src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { AuthModule } from './auth/auth.module';
+import { UsersModule } from './users/users.module';
 
 @Module({
-  imports: [AuthModule],
+  imports: [UsersModule],
   controllers: [],
   providers: [],
 })
