+++ server/users-service/src/users/users.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { UsersService } from './users.service';
+
+@Controller()
+export class UsersController {
+  constructor(private readonly usersService: UsersService) {}
+}
