+++ server/users-service/src/users/users.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthService } from './auth.service';
+import { UsersService } from './users.service';
 
-describe('AuthService', () => {
-  let service: AuthService;
+describe('UsersService', () => {
+  let service: UsersService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [AuthService],
+      providers: [UsersService],
     }).compile();
 
-    service = module.get<AuthService>(AuthService);
+    service = module.get<UsersService>(UsersService);
   });
 
   it('should be defined', () => {
