+++ server/users-service/src/users/users.service.ts
@@ -1,4 +1,4 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AuthService {}
+export class UsersService {}
